Add files via upload
Document Object Model (DOM)

Purpose

Few Web sites are static.  Web sites typically have active elements that appear or disappear or change in other ways depending on the interaction with the user.  At the heart of many of these activities is the Document Object Model (DOM).  In this assignment, you will use JavaScript to interact with the DOM of a Web site.

Assignment Instructions

As a programmer for JScript, Inc., you are part of a team working on a Web site for IT support requests.  The site has many parts and must interact with the user in many ways.  As JavaScript expert on the development team, you have been given the task of writing the JavaScript code that will use the DOM to change elements on the pages and react to DOM events.

Following are the requirements for your JavaScript coding.  For each requirement, you will prepare a short description of the code you will write and prepare an example of how the code will look in the HTML for the page.  Prepare any required element or tag names for your code and document them appropriately.  Your code must use the HTML DOM, and you need only to prepare the code for the DOM interaction, not other support functionality that would lead up to the requirements.

When the user logs into their support account, use a hidden element on the page to display “Hello” + the name of the user.
If the user’s attempt to log in is incorrect, display an error message without reloading the page, and clear the id and password fields.
Depending on the type of support ticket selected from a dropdown list (computer, software, network), display different form fields.
Write one more useful example of the use of JavaScript with DOM for this site. Describe the purpose and the action and prepare an example of the JavaScript in the HTML code.